% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espa_creds.R
\name{espa_creds}
\alias{espa_creds}
\title{Input and Save esap-api Credentials}
\usage{
espa_creds(username, password)
}
\arguments{
\item{username}{espa account's username}

\item{password}{espa account's password corresponding to the username}
}
\value{
NULL. Just saves the username and password in .Reviron
}
\description{
Save the ESPA login credentials as global environment to be used in other functions in rLandsat. This is a pre-requisite for running any of the other functions requiring espa-api
}
\details{
If you do not have an account with espa, please create one here: https://ers.cr.usgs.gov/register
}
\examples{
# set the espa credentials to be used by other functions
espa_creds(username = "your_espaname", password = "secret_password")
}
