% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat_download.R
\name{landsat_download}
\alias{landsat_download}
\title{Downlaod Landsat Files from URL}
\usage{
landsat_download(download_url, dest_file, entity_id = NULL,
  folder_wise = FALSE)
}
\arguments{
\item{download_url}{vector of urls to be downloaded}

\item{dest_file}{the destination folder where the files are to be downloaded}

\item{entity_id}{product id correspoding to the urls if the downloads need to be in folder wise for AWS links}

\item{folder_wise}{if the downloads need to be in folder wise for AWS links}
}
\value{
vector of failed urls
}
\description{
Downlaod Landsat Files from URL
}
\details{
Caution: use entity_id and folder_wise ONLY in case of downloading the individual bands (like from AWS). Get the downlaod urls from espa functions in this library. View the demo
}
\examples{
# save files to the destination path provided
## returns the index of the failed urls
landsat_download("https://edclpdsftp.cr.usgs.gov/orders/order_id1.tar.gz", dest_file = tempdir())

}
