% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat_urls.R
\name{landsat_urls}
\alias{landsat_urls}
\title{Getting Landsat8 Raw Files URL from html.index Files}
\usage{
landsat_urls(index_html_url, TIFtxtOnly = TRUE, band = NULL,
  dest_file = NULL, scrape = FALSE)
}
\arguments{
\item{index_html_url}{index.html url for which download urls are needed}

\item{TIFtxtOnly}{logical. To download only the TIF and txt files or all files}

\item{band}{vector for required bands number}

\item{dest_file}{the destination location}

\item{scrape}{if the urls should be scraped from index.html or created using the pattern}
}
\value{
dataframe with download urls
}
\description{
Getting Landsat8 Raw Files URL from html.index Files
}
\examples{
index_url = "https://s3-us-west-2.amazonaws.com/landsat-pds/c1/L8/148/047/LC08_L1TP_148047_20170404_20170414_01_T1/index.html"
temp = landsat_urls(index_url, dest_file = getwd())
}
