% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_rowpath.R
\docType{data}
\name{world_rowpath}
\alias{world_rowpath}
\title{World - Country wise row path mapping}
\format{An object of class dataframe}
\usage{
data("world_rowpath")
}
\description{
World - Country wise row path mapping
}
\examples{
data("world_rowpath")
}
\keyword{datasets}
